# services/auth-api/openapi.yaml
openapi: 3.0.3
info:
  title: Auth API
  description: Service for handling user authentication, token generation, and user info.
  version: 1.0.0
paths:
  /healthz:
    get:
      summary: Health Check
      tags:
        - Health
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /login:
    post:
      summary: User Login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful Login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /refresh:
    post:
      summary: Refresh Token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Successful Refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
  /me:
    get:
      summary: Get Current User
      tags:
        - User
      security:
        - OAuth2PasswordBearer: []
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Unauthorized

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        ok:
          type: boolean
          default: true
      required:
        - ok
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
      required:
        - username
        - password
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: "bearer"
      required:
        - access_token
        - refresh_token
        - token_type
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required:
        - refresh_token
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
      required:
        - id
        - username
        - email
        - roles
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          tokenUrl: /login
          scopes: {}